start authorisation

+++ Backend/src/main/java/msc/HME/controller/UserController.java
@@ -1,5 +1,6 @@
 package msc.HME.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
 import msc.HME.binding.User;
 import msc.HME.service.UserService;
 import org.springframework.dao.DataAccessException;
@@ -22,7 +23,10 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Object> findUser(@PathVariable String id) {
+    public ResponseEntity<Object> findUser(@PathVariable String id, HttpServletRequest request) {
+        if (!userService.checkAuth(request)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized request");
+        }
         try {
             User result = userService.getUser(id);
             return ResponseEntity.status(HttpStatus.OK).body(result);
@@ -34,7 +38,10 @@ public class UserController {
     }
 
     @GetMapping("/{id}/{resource}") // resource options: s, f, a, w
-    public ResponseEntity<Object> findResource(@PathVariable String id, @PathVariable String resource) {
+    public ResponseEntity<Object> findResource(@PathVariable String id, @PathVariable String resource, HttpServletRequest request) {
+        if (!userService.checkAuth(request)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized request");
+        }
         try {
             if (Objects.equals(resource, "s")) {
                 Object result = userService.findSearches(id);
@@ -61,7 +68,7 @@ public class UserController {
     }
 
     @PostMapping("/new/{id}/{resource}")
-    public ResponseEntity<Object> addResource(@PathVariable String id, @PathVariable String resource,
+    public ResponseEntity<Object> addResource(@PathVariable String id, @PathVariable String resource, HttpServletRequest request,
                                               @RequestParam(required = false ) String searchString,
                                               @RequestParam(required = false ) String propertyId,
                                               @RequestParam(required = false ) String message,
@@ -73,6 +80,9 @@ public class UserController {
                                               @RequestParam(required = false ) String transportation,
                                               @RequestParam(required = false ) String emergency)
     {
+        if (!userService.checkAuth(request)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized request");
+        }
         try {
             if (Objects.equals(resource, "s") && !searchString.isBlank()) {
                 userService.addSearch(id, searchString);
@@ -99,7 +109,10 @@ public class UserController {
     }
 
     @PatchMapping("/update/{id}/email")
-    public ResponseEntity<Object> updateEmail(@PathVariable String id, @RequestParam String email) {
+    public ResponseEntity<Object> updateEmail(@PathVariable String id, @RequestParam String email, HttpServletRequest request) {
+        if (!userService.checkAuth(request)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized request");
+        }
         try {
             return userService.updateEmail(id, email);
         } catch (EmptyResultDataAccessException e) {
@@ -110,7 +123,7 @@ public class UserController {
     }
 
     @PatchMapping("/update/{id}/{resource}") // s, w.
-    public ResponseEntity<Object> updateResources(@PathVariable String id, @PathVariable String resource,
+    public ResponseEntity<Object> updateResources(@PathVariable String id, @PathVariable String resource, HttpServletRequest request,
                                                   @RequestParam(required = false) String searchString,
                                                   @RequestParam(required = false) String newSearchString,
                                                   @RequestParam(required = false ) String entertainment,
@@ -120,6 +133,9 @@ public class UserController {
                                                   @RequestParam(required = false ) String financial,
                                                   @RequestParam(required = false ) String transportation,
                                                   @RequestParam(required = false ) String emergency) {
+        if (!userService.checkAuth(request)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized request");
+        }
         try {
             if (Objects.equals(resource, "s") && searchString != null && newSearchString != null) {
                 userService.addSearch(id, newSearchString);
@@ -141,9 +157,12 @@ public class UserController {
     }
 
     @DeleteMapping("/remove/{id}/{resource}") // s f w
-    public ResponseEntity<Object> removeResource(@PathVariable String id, @PathVariable String resource,
+    public ResponseEntity<Object> removeResource(@PathVariable String id, @PathVariable String resource, HttpServletRequest request,
                                                  @RequestParam(required = false ) String searchString,
                                                  @RequestParam(required = false ) String propertyId) {
+        if (!userService.checkAuth(request)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized request");
+        }
         try {
             if (Objects.equals(resource, "s") && !searchString.isBlank()) {
                 userService.removeSearch(id, searchString);
@@ -167,7 +186,10 @@ public class UserController {
     }
 
     @DeleteMapping("/remove/{id}")
-    public Object removeUser(@PathVariable String id) {
+    public Object removeUser(@PathVariable String id, HttpServletRequest request) {
+        if (!userService.checkAuth(request)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized request");
+        }
         try {
             return userService.deleteUser(id);
         } catch (EmptyResultDataAccessException e) {
